+++ examples/emr_launch_functions/app.py
@@ -64,7 +64,8 @@ launch_cluster = emr_chains.NestedStateMachine(
     name='Launch Cluster StateMachine',
     state_machine=launch_function.state_machine,
     input={
-        'ClusterConfigurationOverrides': sfn.TaskInput.from_data_at('$.ClusterConfigurationOverrides').value
+        'ClusterConfigurationOverrides': sfn.TaskInput.from_data_at('$.ClusterConfigurationOverrides').value,
+        'Tags': sfn.TaskInput.from_data_at('$.Tags').value
     },
     fail_chain=fail).chain